+++ clarinet/services/image/__init__.py
@@ -2,6 +2,7 @@
 
 from clarinet.services.image.coco2nii import COCODataset, coco_to_segmentation
 from clarinet.services.image.image import FileType, Image
+from clarinet.services.image.layered_segmentation import LayeredSegmentation
 from clarinet.services.image.segmentation import (
     PropName,
     Segmentation,
@@ -12,6 +13,7 @@ __all__ = [
     "COCODataset",
     "FileType",
     "Image",
+    "LayeredSegmentation",
     "PropName",
     "Segmentation",
     "coco_to_segmentation",